alt add to base cg shouldn't overlap with cg keys because cording is order-agnostic
On alt-j, if you want to jump to CG 1 by using an alt sequence instead of double-tap, if you chord j-alt by accident, you will end up adding CG 1 to CG 3 by accident. The only way to avoid this is to have the alt add to CG 3 be an alt-combination that does not use another CG key. We still prefer pinky because it minimizes doubling for camera setting.